fflag_api/requirements.txt
# Feature Flag API - Requirements
# Core Flask dependencies
flask>=2.3.0
flask-cors>=4.0.0

# Machine Learning Models
catboost>=1.2.0
xgboost>=1.7.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Semantic Search Dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
torch>=2.0.0
transformers>=4.30.0

# Code Parsing
javalang>=0.13.0

# Configuration
configparser>=5.3.0

# API Client
mistralai>=0.1.0

# Utilities
python-dateutil>=2.8.2
